uint everythin

+++ ports/cxd56/common-hal/analogio/AnalogIn.c
@@ -88,7 +88,7 @@ bool common_hal_analogio_analogin_deinited(analogio_analogin_obj_t *self) {
 
 uint16_t common_hal_analogio_analogin_get_value(analogio_analogin_obj_t *self) {
     uint32_t avg_value = 0;
-    for (int i = 0; i < self->sample_size; i++) {
+    for (uint16_t i = 0; i < self->sample_size; i++) {
         int16_t value = 0;
         read(analogin_dev[self->number].fd, &value, sizeof(value));
         avg_value += (uint16_t)32768 + (uint16_t)value;
